@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -41,6 +41,7 @@ dependencies:
   http: ^0.13.3
   dio: ^4.0.0
   pull_to_refresh: ^2.0.0
+  progress_indicators: ^1.0.0
 
 
 
