@@ -1,9 +1,18 @@
+import 'dart:convert';
+
+import 'package:flutter/cupertino.dart';
+import 'package:http/http.dart';
 import 'package:viewerapp/models/postslist_model.dart';
 import 'package:flutter/foundation.dart';
+
+import '../services/web_services.dart';
 
 class HomePageProvider extends ChangeNotifier {
   bool _showSubCategories1 = false;
   bool _showSubCategories2 = false;
+  String responseCode = "";
+  String message = "";
+  late List<Post> posts = [];
   List<String> _subCategories = [
     "Sub1",
     "Sub2",
@@ -16,16 +25,40 @@ class HomePageProvider extends ChangeNotifier {
     "Sub9",
     "Sub10",
   ];
-  int lastButtonIndex = -1;
-  List<bool> _activeCategories = [false,false,false, false, false, false];
 
-  void like(Post post) {
+  Future<bool> fetchPostsList(int pageSize, int nowPage, String orderBy, int category) async {
+
+    try {
+     Response response =  await WebServices.fetchPosts(pageSize, nowPage, orderBy, category);
+
+     if (response.statusCode == 200) {
+          Map<String, dynamic> decodedResponse = json.decode(utf8.decode(response.bodyBytes));
+          PostsResponse postsResponse = PostsResponse.fromJson(decodedResponse);
+          responseCode = postsResponse.code;
+          message = postsResponse.message;
+          posts.addAll(postsResponse.data);
+          notifyListeners();
+          return true;
+        } else {
+          return false;
+        }
+
+    } catch (e) {
+      notifyListeners();
+      return false;
+    }
+  }
+
+  int lastButtonIndex = -1;
+  List<bool> _activeCategories = [false, false, false, false, false, false];
+
+  void bookmark(Post post) {
     print("like pressed");
     post.like = true;
     notifyListeners();
   }
 
-  void unLike(Post post) {
+  void unbookmark(Post post) {
     print("unlike pressed");
     post.like = false;
     notifyListeners();
@@ -33,7 +66,7 @@ class HomePageProvider extends ChangeNotifier {
 
   void fetchSubCategories(String categoryName, int index) {
     print("fetching  categories");
-    for(int  i = 0; i<_activeCategories.length; i++) {
+    for (int i = 0; i < _activeCategories.length; i++) {
       _activeCategories[i] = false;
     }
     if (_subCategories.length > 10) _subCategories.removeAt(10);
@@ -41,24 +74,20 @@ class HomePageProvider extends ChangeNotifier {
 
     print(lastButtonIndex);
     if (lastButtonIndex == index) {
-
-      if(index == 4)
+      if (index == 4)
         _showSubCategories2 = false;
       else
-      _showSubCategories1 = false;
+        _showSubCategories1 = false;
       lastButtonIndex = -1;
-    }
-    else {
+    } else {
       _activeCategories[index] = true;
       lastButtonIndex = index;
-      if(index == 4) {
+      if (index == 4) {
         _showSubCategories2 = true;
         _showSubCategories1 = false;
-      }
-      else{
-    _showSubCategories1 = true;
-    _showSubCategories2 = false;
-
+      } else {
+        _showSubCategories1 = true;
+        _showSubCategories2 = false;
       }
     }
 
@@ -66,9 +95,10 @@ class HomePageProvider extends ChangeNotifier {
   }
 
   bool get showSubCategories1 => _showSubCategories1;
+
   bool get showSubCategories2 => _showSubCategories2;
 
-
   List<String> get subCategories => _subCategories;
+
   List<bool> get activeAcategories => _activeCategories;
 }
